:bug: fix long to int parse error

+++ src/model/Model.java
@@ -65,16 +65,16 @@ public class Model {
             while (keys.hasNext()) {
                 Object key = keys.next();
                 if (patients.get(key) instanceof JSONObject) {
-                    Task task = new Task((int) (Integer.parseInt((String) key) - 1), (int) patients.size());
+                    Task task = new Task((Integer.parseInt((String) key) - 1), (int) patients.size());
 
                     // Creating tasks
                     tasks.add(task);
 
                     // Creating corresponding visits
-                    Visit visit1 = new Visit(( int) (Integer.parseInt((String) key) - 1), task);
-                    Visit visit1Virtual= new Visit(( int) (Integer.parseInt((String) key) - 1 + 2 * patients.size()), task);
-                    Visit visit2 = new Visit(( int) (Integer.parseInt((String) key) - 1 + patients.size()), task);
-                    Visit visit2Virtual = new Visit(( int) (Integer.parseInt((String) key) - 1 + 3 * patients.size()), task);
+                    Visit visit1 = new Visit((Integer.parseInt((String) key) - 1), task);
+                    Visit visit1Virtual= new Visit((Integer.parseInt((String) key) - 1 + 2 * patients.size()), task);
+                    Visit visit2 = new Visit((Integer.parseInt((String) key) - 1 + patients.size()), task);
+                    Visit visit2Virtual = new Visit((Integer.parseInt((String) key) - 1 + 3 * patients.size()), task);
 
                     this.visits.add(visit1);
                     this.visits.add(visit1Virtual);
@@ -101,10 +101,10 @@ public class Model {
         for (int i = 1; i < jsonTasks.size()+1; i++){
             Task task = new Task(i, this.numTasks);
             JSONObject jsonAttributes = (JSONObject) jsonTasks.get(Integer.toString(i));
-            task.setDuration((int) jsonAttributes.get("care_time"));
-            task.setWeight((int) jsonAttributes.get("demand"));
-            task.setStartTime((int) jsonAttributes.get("start_time"));
-            task.setEndTime((int) jsonAttributes.get("end_time"));
+            task.setDuration(( (Long) jsonAttributes.get("care_time")).intValue());
+            task.setWeight(((Long) jsonAttributes.get("demand")).intValue());
+            task.setStartTime(((Long) jsonAttributes.get("start_time")).intValue());
+            task.setEndTime(((Long) jsonAttributes.get("end_time")).intValue());
             Location location = new Location(i, (Long) jsonAttributes.get("x_coord"), (Long) jsonAttributes.get("x_coord"));
             locations.put(i, location);
             task.setLocation(location);
